Just cleanup I think

+++ distribution cleanup/build catalogue.rb
@@ -1,15 +1,17 @@
+root = ARGV[0].to_s
+raise "Usage: build_catalogue.rb <directory to catalogue>" if ARGV.length < 1
+raise "#{root} is not an existing directory." if !Dir.exist?(root)
+
 require_relative("common.rb")
 
 catalogue = {}
-root = try_get_dir_from_argv(message: "Vanilla OneShot directory:")
-
-scan_files(root, "",  -> (full, relative) {
+each_file(root, "", -> (full, relative) {
 	puts relative
 	catalogue[relative] = hash_file(full)
 })
 
-catalogueFile = File.open("catalogue.json", "w+")
-catalogueFile.print(catalogue.to_json)
-catalogueFile.close()
+catalogue_file = File.open("catalogue.json", "w+")
+catalogue_file.print(catalogue.to_json)
+catalogue_file.close()
 
-puts("> Catalogue built!")
+puts "> Catalogue built"
